# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  schemas:
...
    AdminStats:
      type: object
      properties:
        totalAssets:
          type: integer
          format: int64
          description: Total number of assets in the catalogue
        activeAssets:
          type: integer
          format: int64
          description: Number of active assets
        activeTrustFrameworks:
          type: integer
          format: int64
          description: Number of active trust frameworks
        totalUsers:
          type: integer
          format: int64
          description: Total number of registered users
        totalSchemas:
          type: integer
          format: int64
          description: Total number of schemas
        totalParticipants:
          type: integer
          format: int64
          description: Total number of participants
        graphClaimCount:
          type: integer
          format: int64
          description: Number of claims in the graph database
        graphBackend:
          type: string
          description: The graph database backend type
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    GraphDatabaseStatus:
      type: object
      properties:
        activeBackend:
          type: string
          description: Currently active graph database backend
        connected:
          type: boolean
          description: Whether the graph database is reachable
        claimCount:
          type: integer
          format: int64
          description: Number of claims in the graph database
        version:
          type: string
          description: Graph database version or implementation info
        rebuildNeeded:
          type: boolean
          description: |
            True when the active graph store has no claims but RDF assets exist in
            storage — typical after a backend switch or fresh start. The admin UI
            surfaces a rebuild prompt that calls `POST /admin/graph/rebuild`. Non-RDF
            assets are excluded from this check because they do not produce graph
            claims even after a successful rebuild.
        rdfAssetCount:
          type: integer
          format: int64
          description: |
            Number of active RDF assets in storage that would be re-indexed by a
            rebuild. Surfaced alongside `rebuildNeeded` so the UI can show how many
            assets the rebuild will process.
    GraphDatabaseSwitchResult:
      type: object
      properties:
        message:
          type: string
          description: Human-readable status message
...
    KeycloakAdminUrl:
      type: object
      properties:
        url:
          type: string
          description: Keycloak Admin Console URL for the catalogue realm
    OntologyImpactEntry:
      type: object
      description: >-
        One runtime-uploaded ontology and its contribution to each
        registered trust-framework base class.
      properties:
        id:
          type: string
          description: Schema identifier
        name:
          type: string
          description: >-
            Display name for the ontology. Derived from the parsed
            owl:Ontology IRI when present; falls back to the schema id.
        uploadedAt:
          type: string
          format: date-time
          description: When the ontology was uploaded
        contributions:
          type: object
          description: >-
            For each registered base-class name, the number of distinct
            rdfs:subClassOf+ descendants reachable from the base class's primary
            root and additional_roots, aggregated across active
            trust-framework bundles. Base classes with zero contributions are
            omitted. Empty when parseError is true.
          additionalProperties:
            type: integer
            format: int32
        parseError:
          type: boolean
          default: false
          description: >-
            True when the ontology could not be parsed or the subclass
            walk timed out. The contributions map will be empty; the UI
            should render a warning row so the admin does not conclude
            the ontology is safe to ignore.
        parseErrorMessage:
          type: string
          nullable: true
          description: >-
            Human-readable explanation when parseError is true. Null
            otherwise.
      required:
        - id
        - name
        - contributions
        - parseError
    OntologyImpactList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OntologyImpactEntry'
        noActiveBundles:
          type: boolean
          default: false
          description: >-
            True when no trust-framework bundles are active. In that case
            contributions cannot be computed for any ontology; the UI
            should surface a banner explaining that the OWL toggle has no
            base classes to gate.
      required:
        - items
        - noActiveBundles
...
    RebuildStatus:
      type: object
      properties:
        total:
          type: integer
          format: int64
          description: Total number of assets to process
        processed:
          type: integer
          format: int64
          description: Number of assets processed so far
        percentComplete:
          type: integer
          description: Completion percentage (0-100)
        running:
          type: boolean
          description: Whether a rebuild is currently running
        complete:
          type: boolean
          description: Whether the rebuild has completed
        failed:
          type: boolean
          description: Whether the rebuild has failed
        errorMessage:
          type: string
          description: Error message if the rebuild failed
        errors:
          type: integer
          format: int64
          description: Number of assets that failed during processing
        durationMs:
          type: integer
          format: int64
          description: Duration in milliseconds since rebuild started
...
    SchemaModulePatch:
      type: object
      description: >-
        Merge-patch body for PATCH /admin/schema-validation/modules/{type}. At least
        one field must be present. Only the supplied fields are modified; absent fields
        are left unchanged.
      minProperties: 1
      additionalProperties: false
      properties:
        enabled:
          type: boolean
          description: Desired enabled state of the schema validation module.
...
    SchemaValidationModule:
      type: object
      properties:
        type:
          type: string
          description: Module type (SHACL, JSON_SCHEMA, XML_SCHEMA, OWL)
        name:
          type: string
          description: Display name
        description:
          type: string
          description: Module description
        schemaCount:
          type: integer
          format: int64
          description: Number of schemas for this module
        enabled:
          type: boolean
          description: Whether this module is active
    SchemaValidationStatus:
      type: object
      properties:
        totalSchemaCount:
          type: integer
          format: int64
          description: Total number of schemas across all modules
        modules:
          type: array
          items:
            $ref: '#/components/schemas/SchemaValidationModule'
...
    TrustFrameworkBaseClassPatch:
      type: object
      description: >-
        Merge-patch body for PATCH /admin/trust-frameworks/{bundleId}/base-classes/{baseClassName}.
        At least one field must be present. Only the supplied fields are modified;
        absent fields are left unchanged.
      minProperties: 1
      additionalProperties: false
      properties:
        enabled:
          type: boolean
          description: Desired enabled state of the base class within the bundle.
    TrustFrameworkBundleEffectiveConfig:
      type: object
      description: >-
        Effective compliance configuration of a bundle after applying any per-bundle
        overrides on top of the YAML baseline. Resolved by the catalogue server at
        request time; read-only.
      properties:
        clientType:
          type: string
        serviceUrl:
          type: string
        compliancePath:
          type: string
        apiVersion:
          type: string
        timeoutSeconds:
          type: integer
        trustAnchorUrl:
          type: string
    TrustFrameworkBundleEntry:
      type: object
      properties:
        id:
          type: string
          description: Bundle profile ID (e.g. gaia-x-2511)
        baseClasses:
          type: object
          additionalProperties:
            type: boolean
          description: Per-base-class enabled state for this bundle, keyed by base-class name.
        effectiveConfig:
          $ref: '#/components/schemas/TrustFrameworkBundleEffectiveConfig'
        overriddenFields:
          type: array
          description: >-
            Names of effective-config fields whose value currently comes from a
            persisted override (DB row) rather than the bundle YAML. Allows the
            UI to distinguish operator-set values from YAML defaults.
          items:
            type: string
    TrustFrameworkEntry:
      type: object
      properties:
        id:
          type: string
          description: Trust framework identifier
        name:
          type: string
          description: Display name
        enabled:
          type: boolean
          description: Whether this framework is active
        bundles:
          type: array
          items:
            $ref: '#/components/schemas/TrustFrameworkBundleEntry'
          description: Per-bundle base-class states for this trust framework family.
    TrustFrameworkPatch:
      type: object
      description: >-
        Merge-patch body for PATCH /admin/trust-frameworks/{id}. At least one field
        must be present. Only the supplied fields are modified; absent fields are
        left unchanged.
      minProperties: 1
      additionalProperties: false
      properties:
        enabled:
          type: boolean
          description: Desired enabled state of the trust framework family.
    TrustFrameworkPublicEntry:
      type: object
      properties:
        id:
          type: string
          description: Family ID (e.g. "gaia-x", "mock")
        name:
          type: string
          description: Human-readable family name
        profiles:
          type: array
          items:
            type: string
          description: Profile IDs registered for this family (e.g. ["gaia-x-2511"])
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /admin/me:
    get:
      tags:
        - Admin
      summary: Check if the current user has admin access
      operationId: checkAdminAccess
      security:
        - jwt: []
      responses:
        '200':
          description: User has admin access
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/stats:
    get:
      tags:
        - Admin
      summary: Get dashboard statistics
      operationId: getAdminStats
      security:
        - jwt: []
      responses:
        '200':
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /admin/keycloak-url:
    get:
      tags:
        - Admin
      summary: Get the Keycloak Admin Console URL for embedding
      operationId: getKeycloakAdminUrl
      security:
        - jwt: []
      responses:
        '200':
          description: Keycloak Admin Console URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeycloakAdminUrl'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks:
    get:
      tags:
        - AdminTrustFrameworks
      summary: List all registered trust frameworks
      operationId: getTrustFrameworks
      security:
        - jwt: []
      responses:
        '200':
          description: List of trust frameworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrustFrameworkEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks/{id}:
    patch:
      tags:
        - AdminTrustFrameworks
      summary: Partially update a trust framework
      description: >-
        Applies a merge-patch to the identified trust framework. Only the fields
        present in the request body are modified; absent fields are left unchanged.
        At least one field must be supplied (minProperties: 1). Unknown fields are
        rejected (additionalProperties: false).
      operationId: patchTrustFramework
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/TrustFrameworkPatch'
      responses:
        '200':
          description: Trust framework updated
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks/bundles/{bundleId}:
    patch:
      tags:
        - AdminTrustFrameworks
      summary: Merge-patch external client identifiers of a trust-framework bundle
      description: >-
        Merge-patch semantics (RFC 7396): only fields present in the body are modified.
        Overrides the per-bundle compliance endpoint URL, API version, timeout,
        client implementation key, and trust-anchor list URL. The bundle YAML remains
        the baseline; absent override fields fall back to the YAML value. The change
        takes effect on the next compliance call against the bundle — no restart is
        required. At least one field must be present (minProperties: 1).
      operationId: patchTrustFrameworkBundleConfig
      security:
        - jwt: [ ]
      parameters:
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              type: object
              minProperties: 1
              additionalProperties: true
              description: >-
                Free-form merge-patch object. Recognised keys are clientType,
                serviceUrl, compliancePath, apiVersion, timeoutSeconds, trustAnchorUrl.
                Setting a key's value to a non-null primitive overrides the bundle
                YAML for that field. Setting a key's value to JSON null clears the
                override for that field (the YAML value is used again). Omitting a
                key leaves any existing override or YAML value unchanged. At least
                one key must be present. serviceUrl and trustAnchorUrl must be
                https:// URLs with a public, non-reserved host; other values are
                rejected with 400.
      responses:
        '200':
          description: Bundle configuration updated
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - AdminTrustFrameworks
      summary: Clear all external-identifier overrides for a trust-framework bundle
      description: >-
        Removes any persisted overrides for the given bundle, restoring the bundle YAML
        as the sole source of compliance configuration. Idempotent — a bundle without
        an override row still returns 200. Use when an operator wants to revert all
        runtime adjustments at once; for partial reverts, PATCH with the desired
        fields instead.
      operationId: deleteTrustFrameworkBundleConfig
      security:
        - jwt: [ ]
      parameters:
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bundle overrides cleared (or already absent)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks/{bundleId}/base-classes/{baseClassName}:
    patch:
      tags:
        - AdminTrustFrameworks
      summary: Merge-patch a single base class within a trust-framework bundle
      description: >-
        Merge-patch semantics (RFC 7396): only fields present in the body are modified.
        At least one field must be present.
      operationId: patchTrustFrameworkBaseClass
      security:
        - jwt: []
      parameters:
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
        - name: baseClassName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/TrustFrameworkBaseClassPatch'
      responses:
        '200':
          description: Base-class state updated
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/schema-validation:
    get:
      tags:
        - AdminSchemaValidation
      summary: Get schema validation module status
      operationId: getSchemaValidationStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Schema validation modules with status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaValidationStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/schema-validation/modules/{type}:
    patch:
      tags:
        - AdminSchemaValidation
      summary: Partially update a schema validation module
      description: >-
        Applies a merge-patch to the identified schema validation module. Only the
        fields present in the request body are modified. Disabling a module makes
        affected endpoints return `400 module_disabled:<MODULE>` (see `/assets/validate`
        and `/verification`). Storage of schemas is unaffected — stored schemas remain
        queryable through `/schemas/**` regardless of toggle state.


        **Scope of the `OWL` module toggle:** it controls only whether
        *runtime-uploaded* ontology extensions contribute their subclasses to the
        active trust-framework base classes. The deploy-time
        `verification.semantics` flag is a separate gate with different effects;
        see the arc42 chapter "Verification toggle interactions" for how they
        compose.
      operationId: patchSchemaModule
      security:
        - jwt: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum:
              - SHACL
              - JSON_SCHEMA
              - XML_SCHEMA
              - OWL
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/SchemaModulePatch'
      responses:
        '200':
          description: Module updated
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/schema-validation/ontologies:
    get:
      tags:
        - AdminSchemaValidation
      summary: List uploaded ontologies and their subclass contributions to registered base classes
      description: >-
        Returns the runtime-uploaded ontologies currently consulted by the OWL
        module when it is enabled, with the number of subclasses each ontology
        contributes to each registered trust-framework base class (Participant,
        ServiceOffering, Resource for the Gaia-X 2511 bundle). Read-only —
        admins toggle the OWL module as a whole on
        `PATCH /admin/schema-validation/modules/{type}`. This endpoint surfaces
        what that toggle is gating.
      operationId: getOntologyImpact
      security:
        - jwt: []
      responses:
        '200':
          description: List of ontologies with per-base-class subclass counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyImpactList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/graph-database:
    get:
      tags:
        - AdminGraph
      summary: Get graph database status
      operationId: getGraphDatabaseStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Graph database status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDatabaseStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/graph-database/switch:
    post:
      tags:
        - AdminGraph
      summary: Switch the graph database backend (requires restart)
      operationId: switchGraphDatabase
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backend:
                  type: string
                  description: Target backend (NEO4J, FUSEKI, NONE)
              required:
                - backend
      responses:
        '200':
          description: Switch request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphDatabaseSwitchResult'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/graph/rebuild:
    post:
      tags:
        - AdminGraph
      summary: Trigger an asynchronous graph rebuild
      operationId: triggerGraphRebuild
      security:
        - jwt: []
      responses:
        '202':
          description: Graph rebuild started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebuildStatus'
        '409':
          description: A graph rebuild is already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebuildStatus'
        '503':
          description: Graph store is disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/graph/rebuild/status:
    get:
      tags:
        - AdminGraph
      summary: Get the current graph rebuild status
      operationId: getGraphRebuildStatus
      security:
        - jwt: []
      responses:
        '200':
          description: Current rebuild status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebuildStatus'
...
  /trust-frameworks:
    get:
      tags:
        - TrustFrameworks
      summary: List enabled trust frameworks and their profiles (public discovery)
      operationId: getTrustFrameworksPublic
      security:
        - jwt: [ ]
      responses:
        '200':
          description: List of enabled trust framework families with their profile IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrustFrameworkPublicEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...